Созданы модули по созданию массива объектов и функции
